Separated out config specific function
